fix: ProductCacheServiceTest시 redis 의존성 해결

+++ src/test/java/com/example/store/service/ProductCacheServiceTest.java
@@ -70,9 +70,10 @@ public class ProductCacheServiceTest {
     }
 
     private void clearRedis() {
-        redisTemplate.getConnectionFactory().getConnection().flushDb();
+        if (redisTemplate.getConnectionFactory() != null) {
+            redisTemplate.getConnectionFactory().getConnection().flushDb();
+        }
     }
-
     private void clearDatabase() {
         tagRepository.deleteAll();
     }
